moved map controller to separate directory
