+++ app/index.html
@@ -28,7 +28,7 @@
             </ul>
         </div>
     </div>
-    <main data-ng-controller="SimpleMapController" data-ng-init="getAll()">
+    <main data-ng-controller="MapController" data-ng-init="getAll()">
     <div id="main">
         <div class="header">
             <h1>Seattle Police Report Map</h1>
