+++ app/scripts/map/map.js
@@ -1,0 +1,3 @@
+module.exports = function(app) {
+  require('./controllers/map_controller')(app);
+};
